Vizia GUI refactoring
