+++ src/gui_vizia/components/sext.rs
@@ -2,7 +2,7 @@
 use crate::{
     common::{Component, ViziaComponent},
     components::Sext,
-    gui_vizia::{popup::NewPopup, tooltip::new_component_tooltip},
+    gui_vizia::{gui_components::popup::NewPopup, tooltip::new_component_tooltip},
 };
 
 use vizia::{
