+++ src/gui_vizia/mod.rs
@@ -1,12 +1,11 @@
-mod grid;
 mod gui;
 pub mod hover;
 mod keymap;
-mod menu;
-pub mod popup;
 pub mod tooltip;
 mod transport;
 #[cfg(feature = "components")]
 pub mod components;
+pub mod gui_components;
+mod views;
 
 pub use gui::*;
