+++ src/gui_vizia/views.rs
@@ -1,0 +1,1 @@
+pub mod simulator;
